Feat: ItemOverview props 수정
