--- src/app/my-info/(account)/orders/_components/Orders/OrderItemList/OrderItem/ItemOverview/ItemOverview.module.scss
@@ -1,22 +1,0 @@
-.item {
-  display: flex;
-  align-items: center;
-  gap: 2rem;
-
-  .item-text {
-    @include flex-column(0.4rem);
-    @include pretendard-16-500;
-
-    color: $black;
-
-    .item-name {
-      @include pretendard-14-400;
-
-      color: $gray-50;
-    }
-  }
-
-  .item-option {
-    max-width: 46.2rem;
-  }
-}
