+++ src/app/my-info/(account)/orders/_components/Orders/OrderItemList/OrderItem/OrderItem.tsx
@@ -2,7 +2,7 @@ import classNames from 'classnames/bind';
 
 import { getOrderStatusDescription } from '@/libs/getOrderStatusDescriptions';
 import type { OrderItem as OrderItemT, OrderStatus } from '@/types/orderType';
-import ItemOverview from './ItemOverview/ItemOverview';
+import ItemOverview from '../../../../../../../../components/ItemOverview/ItemOverview';
 import OrderItemButton from './OrderItemButton/OrderItemButton';
 
 import styles from './OrderItem.module.scss';
@@ -22,19 +22,6 @@ export default function OrderItem({ orderItem, confirmationDate, orderStatus }: 
     <div className={cn('order-item')}>
       <div className={cn('item')}>
         <ItemOverview item={orderItem} />
-        {/* <Image
-          src={orderItem.productImgUrl}
-          alt={orderItem.productName}
-          width={107}
-          height={107}
-          placeholder={IMAGE_BLUR.placeholder}
-          blurDataURL={IMAGE_BLUR.blurDataURL}
-        />
-        <div className={cn('item-image')}>
-          <p>{orderItem.productName}</p>
-          <p>{orderItem.switchOption}</p>
-          <CustomOption customData={orderItem.switchOption} />
-        </div> */}
       </div>
       <div className={cn('order-status')}>
         <p>{confirmationDate}</p>
