+++ src/components/index.ts
@@ -19,4 +19,4 @@ export { default as ImageInput } from './WriteEditModal/ImageInput';
 export { default as Button } from './Buttons/Button/Button';
 export { default as ScrollUpButton } from './Buttons/ScrollUpButton/ScrollUpButton';
 
-// export { default as ItemOverview } from './ItemOverview/ItemOverview';
+export { default as ItemOverview } from './ItemOverview/ItemOverview';
